Moved SubscribeCommands into its own file
