+++ src/main/main.cpp
@@ -10,6 +10,8 @@
 
 #include "AppTask.h"
 #include "BindingHandler.h"
+#include "ClusterCommands.h"
+#include "SubscribeCommands.h"
 #include "esp_log.h"
 #include "esp_spi_flash.h"
 #include "esp_system.h"
@@ -17,14 +19,13 @@
 #include "freertos/task.h"
 #include "nvs_flash.h"
 #include "shell_extension/launch.h"
-#include "ClusterCommands.h"
 
 #include <app/server/OnboardingCodesUtil.h>
 
 using namespace ::chip;
 using namespace ::chip::DeviceManager;
 
-static const char *TAG = "light-switch-app";
+static const char * TAG = "light-switch-app";
 
 static AppDeviceCallbacks EchoCallbacks;
 
@@ -55,9 +56,10 @@ extern "C" void app_main()
 #if CONFIG_ENABLE_CHIP_SHELL
     chip::LaunchShell();
     shell::RegisterClusterCommands();
+    shell::RegisterSubscribeCommands();
 #endif // CONFIG_ENABLE_CHIP_SHELL
 
-    CHIPDeviceManager &deviceMgr = CHIPDeviceManager::GetInstance();
+    CHIPDeviceManager & deviceMgr = CHIPDeviceManager::GetInstance();
 
     CHIP_ERROR error = deviceMgr.Init(&EchoCallbacks);
     if (error != CHIP_NO_ERROR)
